refactor: move models to root of `domain` layer
