+++ src/main/java/io/pakland/mdas/githubstats/domain/Commit.java
@@ -1,4 +1,4 @@
-package io.pakland.mdas.githubstats.domain.model;
+package io.pakland.mdas.githubstats.domain;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
