+++ src/main/java/io/pakland/mdas/githubstats/domain/User.java
@@ -1,4 +1,4 @@
-package io.pakland.mdas.githubstats.domain.model;
+package io.pakland.mdas.githubstats.domain;
 
 import javax.persistence.*;
 
@@ -13,61 +13,62 @@ import java.util.List;
 @NoArgsConstructor
 @ToString
 @Entity
-@Table(name = "pull_request")
-public class PullRequest {
+@Table(name = "user")
+public class User {
 
   @Id
   @Column(updatable = false, nullable = false)
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
 
+  private String login;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  private Team team;
+
   @OneToMany(
-    mappedBy = "pullRequest",
+    mappedBy = "user",
     cascade = CascadeType.ALL,
     orphanRemoval = true
   )
   private List<UserReview> userReviews = new ArrayList<>();
 
   @OneToMany(
-    mappedBy = "pullRequest",
+    mappedBy = "user",
     cascade = CascadeType.ALL,
     orphanRemoval = true
   )
   private List<Commit> commits = new ArrayList<>();
 
-  @ManyToOne(fetch = FetchType.LAZY)
-  private Repository repository;
-
   public void addReview(UserReview userReview) {
     userReviews.add(userReview);
-    userReview.setPullRequest(this);
+    userReview.setUser(this);
   }
 
   public void removeReview(UserReview userReview) {
     userReviews.remove(userReview);
-    userReview.setPullRequest(null);
+    userReview.setUser(null);
   }
 
   public void addCommit(Commit commit) {
     commits.add(commit);
-    commit.setPullRequest(this);
+    commit.setUser(this);
   }
 
   public void removeCommit(Commit commit) {
     commits.remove(commit);
-    commit.setPullRequest(null);
+    commit.setUser(null);
   }
 
   @Override
   public boolean equals(Object o) {
     if (this == o) return true;
-    if (!(o instanceof PullRequest)) return false;
-    return id != null && id.equals(((PullRequest) o).getId());
+    if (!(o instanceof User )) return false;
+    return id != null && id.equals(((User) o).getId());
   }
 
   @Override
   public int hashCode() {
     return getClass().hashCode();
   }
-
 }
