+++ src/main/java/io/pakland/mdas/githubstats/domain/repository/CommentRepository.java
@@ -1,6 +1,6 @@
 package io.pakland.mdas.githubstats.domain.repository;
 
-import io.pakland.mdas.githubstats.domain.model.Comment;
+import io.pakland.mdas.githubstats.domain.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
