+++ src/main/java/io/pakland/mdas/githubstats/domain/repository/CommitRepository.java
@@ -1,6 +1,6 @@
 package io.pakland.mdas.githubstats.domain.repository;
 
-import io.pakland.mdas.githubstats.domain.model.Commit;
+import io.pakland.mdas.githubstats.domain.Commit;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
