+++ src/main/java/io/pakland/mdas/githubstats/domain/repository/HistoricQueriesRepository.java
@@ -1,6 +1,6 @@
 package io.pakland.mdas.githubstats.domain.repository;
 
-import io.pakland.mdas.githubstats.domain.model.HistoricQueries;
+import io.pakland.mdas.githubstats.domain.HistoricQueries;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface HistoricQueriesRepository extends JpaRepository<HistoricQueries, Long> {
