+++ src/main/java/io/pakland/mdas/githubstats/domain/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package io.pakland.mdas.githubstats.domain.repository;
 
-import io.pakland.mdas.githubstats.domain.model.User;
+import io.pakland.mdas.githubstats.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
